Minor Tweaks!
This PR permits the taking of the MCU Augment for anyone the company has already trusted with body armour, at the normal points cost. As such, it's available to Security and Command.

It also de-restricts the Tool-Arm augment, which makes it a competitor for limited loadout points and may help draw people away from the combat skill boosting augs by providing a viable alternative.

It also corrects an oversight that led to pistols mags that you could get from cargo not actually fitting in the pistols you could get from cargo.